Move to using XModem to download images to the programmer.
Major update. Instead of requiring the ROM image to be on an SD card, it is now downloaded via XModem during a write or verify operation.
Since there isn't enough local memory to store it, this does mean that a write followed by a verify must download the file twice.